refactor: merge all notifications together
